src/requirements.txt
pillow
psycopg2
django
south
django-debug-toolbar
django-pipeline
